Created porting system, externalized events
Created old colored blocks porting system.
Externalized sync events.
